moved styles folder to web root
