+++ web/app/styles/Hero.module.css
@@ -1,14 +1,21 @@
 .hero {
-    background-color: #232946;
-    color: #fff;
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-  
-  .h1 {
-    font-size: 64px;
-  }
+  background-color: #232946;
+  color: #fff;
+  height: 100vh;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  text-align: center;
+}
+
+.h1 {
+  font-size: 64px;
+  margin-bottom: 0; 
+}
+
+.paragraph {
+  font-size: 24px;
+  margin-top: 0;
+}
