+++ web/components-client/Hero.tsx
@@ -1,12 +1,13 @@
-import React from 'react'
+import React from 'react';
 import styles from '../app/styles/Hero.module.css';
 
 const Hero = () => {
   return (
     <div className={styles.hero}>
-        <h1 className={styles.h1}>Quapp</h1>
+      <h1 className={styles.h1}>Quapp</h1>
+      <p className={styles.paragraph}>Lorem ipsum dolor sit amet.</p>
     </div> 
-  )
+  );
 }
 
-export default Hero
+export default Hero;
